Fix NumberOptions.default type. Remove alias from output
